requirements.txt
boto3==1.24.7
botocore==1.27.7
cachelib==0.7.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
Flask==2.1.2
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.0
MarkupSafe==2.1.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.28.0
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.37
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
